Huge update into ESModules
