--- os/web/consts.js
@@ -1,6 +1,0 @@
-
-const GAMES_CACHE_PREFIX = "fscache";
-// @TODO Remove this
-// const GAMES_CACHE_DELIMITER = '___';
-
-const GAME_PATH = "./game/";
