+++ os/web/consts.mjs
@@ -1,0 +1,6 @@
+
+export const GAMES_CACHE_PREFIX = "fscache";
+// @TODO Remove this
+// const GAMES_CACHE_DELIMITER = '___';
+
+export const GAME_PATH = "./game/";
