+++ os/web/index.html
@@ -20,55 +20,10 @@
         </div>
 
         <script
-            type="text/javascript"
-            src="./pako_inflate.min.js"
-            crossorigin="anonymous"
+            type="module"
+            src="./index.mjs"            
         ></script>
-        <script
-            type="text/javascript"
-            src="./fetcher.js"
-            crossorigin="anonymous"
-        ></script>
-        <script
-            type="text/javascript"
-            src="./asyncfetchfs.js"
-            crossorigin="anonymous"
-        ></script>
-        <script
-            type="text/javascript"
-            src="./tar.js"
-            crossorigin="anonymous"
-        ></script>
-        <script
-            type="text/javascript"
-            src="./config.js"
-            crossorigin="anonymous"
-        ></script>
-        <script
-            type="text/javascript"
-            src="./consts.js"
-            crossorigin="anonymous"
-        ></script>
-        <script
-            type="text/javascript"
-            src="./index.js"
-            crossorigin="anonymous"
-        ></script>
-        <script
-            type="text/javascript"
-            src="./fallout2-ce.js"
-            crossorigin="anonymous"
-        ></script>
-        <script
-            type="text/javascript"
-            src="./iniparser.js"
-            crossorigin="anonymous"
-        ></script>
-        <script
-            type="text/javascript"
-            src="./mainmenu.js"
-            crossorigin="anonymous"
-        ></script>
+        
         <script>
             if ("serviceWorker" in navigator) {
                 if (window.location.hostname !== "localhost") {
@@ -76,11 +31,7 @@
                 }
             }
         </script>
-        <script
-            type="text/javascript"
-            src="./onscreen_keyboard.js"
-            crossorigin="anonymous"
-        ></script>
+        
         <script
             type="text/javascript"
             src="./custom.js"
